Use vim 8 configuration
We now load plugins through native support instead of pathogen. We now
switched to a new theme for the new year. We made our comments more
consistent. We also now support pbcopy of gist created in vim
automatically to the clipboard. We also enabled highlighting since the
theme applied seems to remove it.